# swagger.yaml
openapi: 3.1.0
info:
  title: Умное БТИ
  description: MVP платформы для клиентов, исполнителей и администраторов
  version: 0.1.0
paths:
...
  /api/v1/services/{serviceId}:
    get:
      tags:
      - Public
      summary: Get Service
      operationId: get_service_api_v1_services__serviceId__get
      parameters:
      - name: serviceId
        in: path
        required: true
        schema:
          type: integer
          title: Serviceid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/districts/{code}:
    get:
      tags:
      - Public
      summary: Get District
      operationId: get_district_api_v1_districts__code__get
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistrictRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/house-types/{code}:
    get:
      tags:
      - Public
      summary: Get House Type
      operationId: get_house_type_api_v1_house_types__code__get
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseTypeRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/orders:
    get:
      tags:
      - Admin
      summary: Список заказов (админ)
      operationId: list_orders_api_v1_admin_orders_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
                title: Response List Orders Api V1 Admin Orders Get
      security:
      - HTTPBearer: []
  /api/v1/admin/orders/{order_id}:
    get:
      tags:
      - Admin
      summary: Детали заказа (админ)
      operationId: get_order_api_v1_admin_orders__order_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Admin
      summary: Обновление заказа (админ)
      operationId: update_order_api_v1_admin_orders__order_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateOrderRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/orders/{order_id}/assign-executor:
    post:
      tags:
      - Admin
      - Admin
      summary: Assign Executor
      operationId: assign_executor_api_v1_admin_orders__order_id__assign_executor_post
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignExecutorRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/orders/{order_id}/schedule-visit:
    post:
      tags:
      - Admin
      - Admin
      summary: Schedule Visit
      operationId: schedule_visit_api_v1_admin_orders__order_id__schedule_visit_post
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleVisitRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutorCalendarEvent'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Admin
      - Admin
      summary: Update Visit
      operationId: update_visit_api_v1_admin_orders__order_id__schedule_visit_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleVisitUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutorCalendarEvent'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/services:
    get:
      tags:
      - Admin
      summary: List Services
      operationId: list_services_api_v1_admin_services_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ServiceRead'
                type: array
                title: Response List Services Api V1 Admin Services Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - Admin
      summary: Create Service
      operationId: create_service_api_v1_admin_services_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/admin/services/{code}:
    patch:
      tags:
      - Admin
      summary: Update Service
      operationId: update_service_api_v1_admin_services__code__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: integer
          title: Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/departments:
    get:
      tags:
      - Admin
      summary: List Departments
      operationId: list_departments_api_v1_admin_departments_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DepartmentRead'
                type: array
                title: Response List Departments Api V1 Admin Departments Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - Admin
      summary: Create Department
      operationId: create_department_api_v1_admin_departments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/admin/departments/{code}:
    patch:
      tags:
      - Admin
      summary: Update Department
      operationId: update_department_api_v1_admin_departments__code__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/districts:
    post:
      tags:
      - Admin
      summary: Create District
      operationId: create_district_api_v1_admin_districts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistrictCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistrictRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/admin/districts/{code}:
    patch:
      tags:
      - Admin
      summary: Update District
      operationId: update_district_api_v1_admin_districts__code__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistrictUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistrictRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/house-types:
    post:
      tags:
      - Admin
      summary: Create House Type
      operationId: create_house_type_api_v1_admin_house_types_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseTypeCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseTypeRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/admin/house-types/{code}:
    patch:
      tags:
      - Admin
      summary: Update House Type
      operationId: update_house_type_api_v1_admin_house_types__code__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseTypeUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseTypeRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AdminUpdateOrderRequest:
      properties:
        status:
          anyOf:
          - $ref: '#/components/schemas/OrderStatus'
          - type: 'null'
        currentDepartmentCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Currentdepartmentcode
        estimatedPrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Estimatedprice
        totalPrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Totalprice
      type: object
      title: AdminUpdateOrderRequest
...
    AssignExecutorRequest:
      properties:
        executorId:
          type: string
          format: uuid
          title: Executorid
      type: object
      required:
      - executorId
      title: AssignExecutorRequest
...
    CalendarStatus:
      type: string
      enum:
      - PLANNED
      - DONE
      - CANCELLED
      title: CalendarStatus
...
    DepartmentCreate:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - code
      - name
      title: DepartmentCreate
    DepartmentRead:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - code
      - name
      title: DepartmentRead
    DepartmentUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      title: DepartmentUpdate
    DistrictCreate:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        priceCoef:
          anyOf:
          - type: number
          - type: 'null'
          title: Pricecoef
          default: 1.0
      type: object
      required:
      - code
      - name
      title: DistrictCreate
    DistrictRead:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        priceCoef:
          anyOf:
          - type: number
          - type: 'null'
          title: Pricecoef
          default: 1.0
      type: object
      required:
      - code
      - name
      title: DistrictRead
    DistrictUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        priceCoef:
          anyOf:
          - type: number
          - type: 'null'
          title: Pricecoef
      type: object
      title: DistrictUpdate
...
    ExecutorCalendarEvent:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        executorId:
          type: string
          format: uuid
          title: Executorid
        orderId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Orderid
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        startTime:
          type: string
          format: date-time
          title: Starttime
        endTime:
          type: string
          format: date-time
          title: Endtime
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        status:
          $ref: '#/components/schemas/CalendarStatus'
        createdAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Createdat
      type: object
      required:
      - id
      - executorId
      - startTime
      - endTime
      - status
      title: ExecutorCalendarEvent
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HouseTypeCreate:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        priceCoef:
          anyOf:
          - type: number
          - type: 'null'
          title: Pricecoef
          default: 1.0
      type: object
      required:
      - code
      - name
      title: HouseTypeCreate
    HouseTypeRead:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        priceCoef:
          anyOf:
          - type: number
          - type: 'null'
          title: Pricecoef
          default: 1.0
      type: object
      required:
      - code
      - name
      title: HouseTypeRead
    HouseTypeUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        priceCoef:
          anyOf:
          - type: number
          - type: 'null'
          title: Pricecoef
      type: object
      title: HouseTypeUpdate
...
    Order:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        clientId:
          type: string
          format: uuid
          title: Clientid
        serviceCode:
          type: integer
          title: Servicecode
        status:
          $ref: '#/components/schemas/OrderStatus'
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        districtCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Districtcode
        houseTypeCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Housetypecode
        complexity:
          anyOf:
          - type: string
          - type: 'null'
          title: Complexity
        calculatorInput:
          anyOf:
          - type: object
          - type: 'null'
          title: Calculatorinput
        estimatedPrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Estimatedprice
        totalPrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Totalprice
        currentDepartmentCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Currentdepartmentcode
        aiDecisionStatus:
          anyOf:
          - type: string
          - type: 'null'
          title: Aidecisionstatus
        aiDecisionSummary:
          anyOf:
          - type: string
          - type: 'null'
          title: Aidecisionsummary
        plannedVisitAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Plannedvisitat
        completedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completedat
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updatedat
      type: object
      required:
      - id
      - clientId
      - serviceCode
      - status
      - title
      - createdAt
      title: Order
...
    OrderStatus:
      type: string
      enum:
      - DRAFT
      - SUBMITTED
      - EXECUTOR_ASSIGNED
      - VISIT_SCHEDULED
      - DOCUMENTS_IN_PROGRESS
      - COMPLETED
      - CANCELLED
      - REJECTED
      title: OrderStatus
...
    ScheduleVisitRequest:
      properties:
        executorId:
          type: string
          format: uuid
          title: Executorid
        startTime:
          type: string
          format: date-time
          title: Starttime
        endTime:
          type: string
          format: date-time
          title: Endtime
        location:
          type: string
          title: Location
      type: object
      required:
      - executorId
      - startTime
      - endTime
      - location
      title: ScheduleVisitRequest
    ScheduleVisitUpdateRequest:
      properties:
        executorId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Executorid
        startTime:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Starttime
        endTime:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Endtime
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      type: object
      title: ScheduleVisitUpdateRequest
...
    ServiceCreate:
      properties:
        code:
          type: integer
          title: Code
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        departmentCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Departmentcode
        basePrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Baseprice
        baseDurationDays:
          anyOf:
          - type: integer
          - type: 'null'
          title: Basedurationdays
        requiredDocs:
          anyOf:
          - type: object
          - type: 'null'
          title: Requireddocs
        isActive:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Isactive
          default: true
      type: object
      required:
      - code
      - title
      title: ServiceCreate
    ServiceRead:
      properties:
        code:
          type: integer
          title: Code
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        departmentCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Departmentcode
        basePrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Baseprice
        baseDurationDays:
          anyOf:
          - type: integer
          - type: 'null'
          title: Basedurationdays
        requiredDocs:
          anyOf:
          - type: object
          - type: 'null'
          title: Requireddocs
        isActive:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Isactive
          default: true
      type: object
      required:
      - code
      - title
      title: ServiceRead
    ServiceUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        departmentCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Departmentcode
        basePrice:
          anyOf:
          - type: number
          - type: 'null'
          title: Baseprice
        baseDurationDays:
          anyOf:
          - type: integer
          - type: 'null'
          title: Basedurationdays
        requiredDocs:
          anyOf:
          - type: object
          - type: 'null'
          title: Requireddocs
        isActive:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Isactive
      type: object
      title: ServiceUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer